+++ src/comps/Background.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react'
 import { motion, useMotionValue, animate } from "framer-motion"
-import './Background.css'
-import democrats from './assets/democrats.png'
-import republicans from './assets/republicans.png'
+import '../css/Background.css'
+import democrats from '../assets/democrats.png'
+import republicans from '../assets/republicans.png'
 
 function Background() {
     return(
