+++ src/css/Background.css
@@ -3,14 +3,14 @@
     flex-wrap: nowrap;
     flex-direction: row;
     justify-content: center;
-    max-width: 100vw;
+    max-width: 100%;
 }
 .left{
     /*
     background-image: url("./assets/democrats.png");
     background-repeat: no-repeat;*/
     height: 100vh;
-    width: 50vw;
+    width: 50%;
     
 }
 .right{
@@ -19,7 +19,7 @@
     background-repeat: no-repeat; 
     background-position: top right;*/
     height: 100vh;
-    width: 50vw;
+    width: 50%;
     
    
    align-content: right;
